Refactor: remove obsolete Hub and Login pages; add new HubPage and LoginPage implementations
