+++ TMS.App/Views/Login/LoginPage.cs
@@ -1,0 +1,12 @@
+﻿using TMS.App.ViewModels;
+
+namespace TMS.App.Views.Login;
+
+public partial class LoginPage : ContentPage
+{
+	public LoginPage(LoginPageViewModel vm)
+	{
+		InitializeComponent();
+		BindingContext = vm ?? throw new ArgumentNullException(nameof(vm));
+	}
+}
