--- TMS.Domain/DTOs/LinesUpdateDTO.cs
@@ -1,9 +1,0 @@
-namespace TMS.Domain.DTOs;
-
-  public sealed record LinesUpdateDTO
-{
-    public Guid Id { get; init; }
-    public double Latitude { get; init; }
-    public double Longitude { get; init; }
-    public required string LineCode { get; init; }
-}
