+++ TMS.Domain/DTOs/MapLineDTO.cs
@@ -1,0 +1,9 @@
+namespace TMS.Domain.DTOs;
+
+public sealed record MapLinesDTO
+{
+  public Guid Id { get; init; }
+  public double Latitude { get; init; }
+  public double Longitude { get; init; }
+  public required string LineCode { get; init; }
+}
